+++ CHANGELOG.md
@@ -4,7 +4,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-- Added cross-format `StandardProperties` class, useful for metadata handling
+- Added CSRG writer
+- Added TSRG and TSRG2 writer
+- Added JAM reader and writer
+- Added JOBF reader and writer
+- Added Recaf Simple reader and writer
+- Added `OuterClassNameInheritingVisitor`
+- Added `MappingFormat#hasWriter` boolean
+- Added cross-format `StandardProperties` class
 - Allowed Tiny v1 to save arbitrary metadata, not only intermediary counters
 
 ## [0.5.1] - 2023-11-30